Fix Reminder constructor

+++ app/src/main/java/io/ipoli/android/quest/data/Reminder.java
@@ -38,7 +38,7 @@ public class Reminder extends RealmObject {
     public Reminder() {
     }
 
-    public Reminder(int minutesFromStart, int notificationId, int intentId) {
+    public Reminder(long minutesFromStart, int notificationId, int intentId) {
         this.id = IDGenerator.generate();
         this.intentId = intentId;
         this.notificationId = notificationId;
